CLI updates, added ability to query ChromaDB

+++ contextify.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
 """A script to embed files to enable RAG for AI prompts"""
 
+import argparse
 import json
 from typing import Optional
 
@@ -101,8 +102,25 @@ class Contextify():
                 self.chroma_api.add_file(context, str(relative_path), file.file_hash,
                                          file_contents, i, file_contents_as_list_len, embeddings)
 
+    def query(self, query: str, context: str):
+        """Queries the context `context` in ChromaDB database with `query`"""
+        ollama_response = self.ollama_client.embed(
+            model=self.embedding_model,
+            input=query
+        )
+        print(ollama_response)
+        return self.chroma_api.query(context, ollama_response['embeddings'])
+
+# pylint: disable=too-many-locals
 def main():
     """The main method"""
+    parser = argparse.ArgumentParser(description="Demo")
+    parser.add_argument('-q', '--query', help='The query (prompt) to use')
+    parser.add_argument('-c', '--context', help="""Specifies the context to use for a query""")
+    parser.add_argument('-s', '--scan', action='store_true', help="""
+                        Scans the configured directories and populates the ChromaDB collections""")
+    args = parser.parse_args()
+
     config = {}
     with open('config.json', 'r', encoding='utf-8') as file:
         config = json.load(file)
@@ -115,23 +133,27 @@ def main():
 
     contextify = Contextify(ollama_url, embedding_model)
 
-    for context in iter(contexts):
-        roots = context["roots"]
-        include_pattern = context["includePattern"]
-        context_name = context["name"]
+    if args.query is not None and args.context is not None:
+        result = contextify.query(args.query, args.context)
+        print(result)
+    if args.scan:
+        for context in iter(contexts):
+            roots = context["roots"]
+            include_pattern = context["includePattern"]
+            context_name = context["name"]
 
-        print(f"Processing context: {context_name}")
-        print(f"Roots: {roots}")
-        print(f"Include pattern: {include_pattern}")
+            print(f"Processing context: {context_name}")
+            print(f"Roots: {roots}")
+            print(f"Include pattern: {include_pattern}")
 
-        files_dict = contextify.get_new_or_changed_files(roots, context_name, include_pattern)
-        new_files = files_dict['new']
-        updated_files = files_dict['updated']
-        for new_file in iter(new_files):
-            contextify.save_file(new_file, context_name)
-        for updated_file in iter(updated_files):
-            contextify.delete_file(updated_file, context_name)
-            contextify.save_file(updated_file, context_name)
+            files_dict = contextify.get_new_or_changed_files(roots, context_name, include_pattern)
+            new_files = files_dict['new']
+            updated_files = files_dict['updated']
+            for new_file in iter(new_files):
+                contextify.save_file(new_file, context_name)
+                for updated_file in iter(updated_files):
+                    contextify.delete_file(updated_file, context_name)
+                    contextify.save_file(updated_file, context_name)
 
 if __name__ == "__main__":
     main()
